requirements.txt
torch==2.8.0
torchvision==0.23.0
numpy==2.3.2
pillow==11.3.0
scikit-learn==1.7.1
matplotlib==3.10.5
pandas==2.3.2